attempt to display current time when playing (it's not really working...)

+++ Cinder/Source/voxel-test-4-bones/include/CinderApp.h
@@ -180,5 +180,5 @@ protected:
 
 	// Animation
 	bool play;
-	int previousTime;
+	float previousTime;
 };
